+++ example/lib/main.dart
@@ -17,17 +17,17 @@ class MyApp extends StatefulWidget {
 
 class SwephTestData {
   final String swephVersion;
-  final double moonLongitude;
-  final double starDistance;
-  final String heavenlyBodyName;
-  final double houseSystemAscmc;
+  final String moonPosition;
+  final String starDistance;
+  final String asteroidName;
+  final String houseSystemAscmc;
   final String chironPosition;
 
   SwephTestData(Sweph sweph)
       : swephVersion = getVersion(sweph),
-        moonLongitude = getMoonLongitude(sweph),
+        moonPosition = getMoonLongitude(sweph),
         starDistance = getStarName(sweph),
-        heavenlyBodyName = getAstroidName(sweph),
+        asteroidName = getAstroidName(sweph),
         houseSystemAscmc = getHouseSystemAscmc(sweph),
         chironPosition = getChironPosition(sweph);
 
@@ -35,28 +35,29 @@ class SwephTestData {
     return sweph.swe_version();
   }
 
-  static double getMoonLongitude(Sweph sweph) {
+  static String getMoonLongitude(Sweph sweph) {
     final jd =
         sweph.swe_julday(2022, 6, 29, (2 + 52 / 60), CalendarType.SE_GREG_CAL);
-    return sweph
-        .swe_calc_ut(jd, HeavenlyBody.SE_MOON, SwephFlag.SEFLG_SWIEPH)
-        .longitude;
+    final pos =
+        sweph.swe_calc_ut(jd, HeavenlyBody.SE_MOON, SwephFlag.SEFLG_SWIEPH);
+    return "lat=${pos.latitude.toStringAsFixed(3)} lon=${pos.longitude.toStringAsFixed(3)}";
   }
 
-  static double getStarName(Sweph sweph) {
+  static String getStarName(Sweph sweph) {
     final jd =
         sweph.swe_julday(2022, 6, 29, (2 + 52 / 60), CalendarType.SE_GREG_CAL);
     return sweph
         .swe_fixstar2_ut('Rohini', jd, SwephFlag.SEFLG_SWIEPH)
         .coordinates
-        .distance;
+        .distance
+        .toStringAsFixed(3);
   }
 
   static String getAstroidName(Sweph sweph) {
     return sweph.swe_get_planet_name(HeavenlyBody.SE_AST_OFFSET + 16);
   }
 
-  static double getHouseSystemAscmc(Sweph sweph) {
+  static String getHouseSystemAscmc(Sweph sweph) {
     const year = 1947;
     const month = 8;
     const day = 15;
@@ -70,7 +71,7 @@ class SwephTestData {
     sweph.swe_set_sid_mode(SiderealMode.SE_SIDM_LAHIRI,
         SiderealModeFlag.SE_SIDBIT_NONE, 0.0 /* t0 */, 0.0 /* ayan_t0 */);
     final result = sweph.swe_houses(julday, latitude, longitude, Hsys.P);
-    return result.ascmc[0];
+    return result.ascmc[0].toStringAsFixed(3);
   }
 
   static String getChironPosition(Sweph sweph) {
@@ -80,13 +81,12 @@ class SwephTestData {
     sweph.swe_julday(2022, 6, 29, (2 + 52 / 60), CalendarType.SE_GREG_CAL);
     final pos =
         sweph.swe_calc_ut(jd, HeavenlyBody.SE_CHIRON, SwephFlag.SEFLG_SWIEPH);
-    return "lat=${pos.latitude} lon=${pos.longitude}";
+    return "lat=${pos.latitude.toStringAsFixed(3)} lon=${pos.longitude.toStringAsFixed(3)}";
   }
 }
 
 class _MyAppState extends State<MyApp> {
-  final sweph =
-      Sweph.getInstance(ephePaths: 'assets${Platform.pathSeparator}ephe');
+  final sweph = Sweph();
   late Future<SwephTestData> swephTestData;
 
   @override
@@ -96,17 +96,61 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<SwephTestData> getTestData() async {
-    // Extracts the resource 'assets/files/seas_18.se1' to 'assets/ephe/seas_18.se1'
+    String defaultEphePath =
+        await sweph.useDefaultEpheFiles(); // Extracts included ephe files
+
+    // Extracts the resource included in example app
     await ResourceUtils.extractAssets(
         'assets/files/seas_18.se1', 'assets/ephe/seas_18.se1');
-    return SwephTestData(await sweph);
+    // Use both folders: the default path where included ephe files are extracted and
+    // the ones packaged with the app
+    sweph.swe_set_ephe_path(
+        '$defaultEphePath${Sweph.folderSeparator}assets${Platform.pathSeparator}ephe');
+    return SwephTestData(sweph);
+  }
+
+  void _addText(List<Widget> children, String text) {
+    const textStyle = TextStyle(fontSize: 25);
+    const spacerSmall = SizedBox(height: 10);
+
+    children.add(spacerSmall);
+    children.add(Text(
+      text,
+      style: textStyle,
+      textAlign: TextAlign.center,
+    ));
+  }
+
+  Widget _getContent(BuildContext context, SwephTestData? swephTestData) {
+    List<Widget> children = [
+      Text(
+        'Swiss Ephemeris Exmaple (working dir = ${Directory.current})',
+        style: const TextStyle(fontSize: 30),
+        textAlign: TextAlign.center,
+      ),
+      const SizedBox(height: 10)
+    ];
+
+    if (swephTestData == null) {
+      _addText(children, 'loading...');
+    } else {
+      _addText(children, 'Sweph Version: ${swephTestData.swephVersion}');
+      _addText(children,
+          'Moon position on 2022-06-29 02:52:00 UTC: ${swephTestData.moonPosition}');
+      _addText(children,
+          'Distance of star Rohini: ${swephTestData.starDistance} AU');
+      _addText(children, 'Name of Asteroid 16: ${swephTestData.asteroidName}');
+      _addText(children,
+          'House System ASCMC[0] for custom time: ${swephTestData.houseSystemAscmc}');
+      _addText(
+          children, 'Chriron position now: ${swephTestData.chironPosition}');
+    }
+
+    return Column(children: children);
   }
 
   @override
   Widget build(BuildContext context) {
-    const textStyle = TextStyle(fontSize: 25);
-    const spacerSmall = SizedBox(height: 10);
-
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
@@ -116,104 +160,12 @@ class _MyAppState extends State<MyApp> {
           child: Center(
             child: Container(
               padding: const EdgeInsets.all(10),
-              child: Column(
-                children: [
-                  Text(
-                    'Dart binding for Swiss Ephemeris (cwd = ${Directory.current})',
-                    style: textStyle,
-                    textAlign: TextAlign.center,
-                  ),
-                  spacerSmall,
-                  FutureBuilder<SwephTestData>(
-                    future: swephTestData,
-                    builder: (BuildContext context,
-                        AsyncSnapshot<SwephTestData> value) {
-                      final sweVersion = (value.hasData)
-                          ? value.data!.swephVersion
-                          : 'loading';
-                      return Text(
-                        'sweph.swe_version = $sweVersion',
-                        style: textStyle,
-                        textAlign: TextAlign.center,
-                      );
-                    },
-                  ),
-                  spacerSmall,
-                  FutureBuilder<SwephTestData>(
-                    future: swephTestData,
-                    builder: (BuildContext context,
-                        AsyncSnapshot<SwephTestData> value) {
-                      final moonLongitude = (value.hasData)
-                          ? value.data!.moonLongitude
-                          : 'loading';
-                      return Text(
-                        'Moon longitude on 2022-06-29 02:52:00 UTC = $moonLongitude',
-                        style: textStyle,
-                        textAlign: TextAlign.center,
-                      );
-                    },
-                  ),
-                  spacerSmall,
-                  FutureBuilder<SwephTestData>(
-                    future: swephTestData,
-                    builder: (BuildContext context,
-                        AsyncSnapshot<SwephTestData> value) {
-                      final displayValue = (value.hasData)
-                          ? value.data!.starDistance
-                          : 'loading';
-                      return Text(
-                        'Distance of star Rohini = $displayValue AU',
-                        style: textStyle,
-                        textAlign: TextAlign.center,
-                      );
-                    },
-                  ),
-                  spacerSmall,
-                  FutureBuilder<SwephTestData>(
-                    future: swephTestData,
-                    builder: (BuildContext context,
-                        AsyncSnapshot<SwephTestData> value) {
-                      final displayValue = (value.hasData)
-                          ? value.data!.heavenlyBodyName
-                          : 'loading';
-                      return Text(
-                        'Name of custom planet = $displayValue',
-                        style: textStyle,
-                        textAlign: TextAlign.center,
-                      );
-                    },
-                  ),
-                  spacerSmall,
-                  FutureBuilder<SwephTestData>(
-                    future: swephTestData,
-                    builder: (BuildContext context,
-                        AsyncSnapshot<SwephTestData> value) {
-                      final displayValue = (value.hasData)
-                          ? value.data!.houseSystemAscmc
-                          : 'loading';
-                      return Text(
-                        'House System ASCMC[0] for custom time = $displayValue AU',
-                        style: textStyle,
-                        textAlign: TextAlign.center,
-                      );
-                    },
-                  ),
-                  spacerSmall,
-                  FutureBuilder<SwephTestData>(
-                    future: swephTestData,
-                    builder: (BuildContext context,
-                        AsyncSnapshot<SwephTestData> value) {
-                      final pos = (value.hasData)
-                          ? value.data!.chironPosition
-                          : 'loading';
-                      return Text(
-                        'Chriron position now = $pos',
-                        style: textStyle,
-                        textAlign: TextAlign.center,
-                      );
-                    },
-                  ),
-                ],
+              child: FutureBuilder<SwephTestData>(
+                future: swephTestData,
+                builder:
+                    (BuildContext context, AsyncSnapshot<SwephTestData> value) {
+                  return _getContent(context, value.data);
+                },
               ),
             ),
           ),